+++ osmomath/math.go
@@ -19,6 +19,16 @@ var (
 	// https://www.wolframalpha.com/input?i=2.718281828459045235360287471352662498&assumption=%22ClashPrefs%22+-%3E+%7B%22Math%22%7D
 	// nolint: unused
 	eulersNumber = MustNewBigDecFromStr("2.718281828459045235360287471352662498")
+
+	TenE9 = NewInt(1_000_000_000)
+	TenE8 = NewInt(100_000_000)
+	TenE7 = NewInt(10_000_000)
+	TenE6 = NewInt(1_000_000)
+	TenE5 = NewInt(100_000)
+	TenE4 = NewInt(10_000)
+	TenE3 = NewInt(1_000)
+	TenE2 = NewInt(100)
+	TenE1 = NewInt(10)
 )
 
 // Returns the internal "power precision".
@@ -202,4 +212,47 @@ func OrderOfMagnitude(num Dec) int {
 	}
 
 	return order
+}
+
+// GetPrecomputeOrderOfMagnitude returns the order of magnitude of the given amount.
+// Uses look up table for precomputed order of magnitudes.
+// CONTRACT: amount must be positive or zero. Panics if not
+func GetPrecomputeOrderOfMagnitude(amount Int) int {
+	if amount.IsNegative() {
+		panic(fmt.Errorf("amount must be positive or zero, was (%s)", amount))
+	}
+
+	if amount.GT(TenE9) {
+		a := amount.Quo(TenE9)
+		return 9 + GetPrecomputeOrderOfMagnitude(a)
+	}
+	if amount.GTE(TenE9) {
+		return 9
+	}
+	if amount.GTE(TenE8) {
+		return 8
+	}
+	if amount.GTE(TenE7) {
+		return 7
+	}
+	if amount.GTE(TenE6) {
+		return 6
+	}
+	if amount.GTE(TenE5) {
+		return 5
+	}
+	if amount.GTE(TenE4) {
+		return 4
+	}
+	if amount.GTE(TenE3) {
+		return 3
+	}
+	if amount.GTE(TenE2) {
+		return 2
+	}
+	if amount.GTE(TenE1) {
+		return 1
+	}
+
+	return 0
 }